requirements.txt
Django>=5.0,<6.0
numpy>=1.26
Pillow>=10.0
opencv-python>=4.9
